fix: Restrict Kubechecks to a Single Namespace for App Watcher
Currently, Kubechecks can monitor all ArgoCD applications and ApplicationSet CRDs across the entire cluster. This is not ideal in scenarios where multiple ArgoCD instances are running on a single cluster, such as for testing purposes.

This change introduces the ability for the operator to specify a namespace for ArgoCD. As a result, Kubechecks will be limited to watching ArgoCD CRDs within the specified namespace only.