remove application set revision check.
applicationsets was checking the manifest was targeting main, however there are cases where targetRevision may be using template e.g. `{{ .values.target }}`
This causes the check to fail and does not perform applicationset checks.

This PR removes the check as this check is only required on the applications not applicationsets.